requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
whitenoise==6.7.0
docopt==0.6.2
idna==3.6
pipreqs==0.4.13
psycopg2-binary==2.9.9
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.0
yarg==0.1.9
pillow==10.2.0
